feat(RoomTime): post and delete functionality, and refactoring
